[REFACTOR] : SampleDataFactory에서 사용하는 예외들의 패키지 이동
같은 패키지로 묶어줌
